+++ resources/js/app.jsx
@@ -2,36 +2,20 @@ import './bootstrap'
 import '../css/app.css'
 import 'react-toastify/dist/ReactToastify.css'
 
-import React from 'react'
+import React, { useEffect } from 'react'
 import ReactDOM from 'react-dom/client'
 import AppRoutes from './routes'
-import axios from 'axios'
-import { useNavigate } from 'react-router-dom'
+import setupAxiosInterceptors from './auth/axiosInterceptor'
 
-window.axios = axios
-window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest'
-window.axios.defaults.baseURL = 'http://task-scheduler.test/'
-const token = localStorage.getItem('token')
-if (token) {
-  axios.defaults.headers.common.Authorization = `Bearer ${token}`
+const App = () => {
+  useEffect(() => {
+    setupAxiosInterceptors()
+  }, [])
+
+  return <AppRoutes />
 }
-const navigate = useNavigate();
-
-axios.interceptors.response.use(
-  response => response,
-  error => {
-    if (error.response?.status === 401) {
-
-      localStorage.removeItem('token')
-      axios.defaults.headers.common.Authorization = 'Bearer'
-
-      navigate('/login')
-    }
-    return Promise.reject(error)
-  }
-)
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
-  <AppRoutes />
+  <App />
 )
